requirements.txt
streamlit
pandas
numpy
openai
plotly
fpdf2
python-dotenv
